requirements.txt
Django==1.8.2
Pillow==2.9.0
django-summernote==0.6.8
gunicorn==19.3.0
python-dateutil==2.4.2
simplejson==3.7.3
sorl-thumbnail==12.3
whitenoise==2.0.2
-e git+https://github.com/ionata/django-singletons.git#egg=django_singletons
pytz==2015.4
mock==1.1.3
psycopg2==2.6.1
django-storages==1.1.8
boto==2.38.0
